creating links ought not be up to the individual templater but a part of the core
